refactor(*): move config file from `examples` to `config` directory
Some package pruners may remove files from exmaples directory. This change will prevent from this.
